CodeFileToolbar renaming to avoid ambiguity
