+++ client/src/components/ide/codeFile/panes/CodeFilePanes.scss
