+++ client/src/components/ide/codeFile/panes/Output.scss
