+++ client/src/components/ide/codeFile/panes/OutputToolbar.scss
